Refactor sentinel subpackage.
This commit contains a large refactor of the existing sentinel
subpackage.
- Separate out logic for dealing with various metadata files to
promote readability and malliablility
- Allow HREFs to be translated to different HREFs for reading, which
will be useful when we need to append SAS tokens (for Azure) or
otherwise use a different HREF to read files as opposed to always
reading from the HREFs the functionlity places in the STAC assets.
- Generate proj metadata