Adds a validate cli command
This is a very simple cli command that runs validate (for Items) or
validate_all (for Collections and Catalogs). There is also an `--only`
flag to only run validate for container objects.

The output is a simple success message, or the exception message and
source. This should hopefully help folks track down validation errors in
existing STAC objects that can be hard to parse from exception output.