feat: remove GDAL Python bindings dependency
This requires refactoring `add_raster_to_item` to use rasterio instead
of GDAL bindings. During this refactor, I discovered that the original
implementation was not correctly handling the top bound of the histogram
calculation, and so the last bin value was one too low in the tests. The
test case was corrected.